requirements.txt
Flask==3.0.0
flask-cors==4.0.0
openai==1.51.2
python-dotenv==1.0.0
